pyproject.toml
[tool.poetry]
name = "notionimportebook"
version = "0.1.0"
description = "import ebook to notions"
authors = ["harumonia <zxjlm233@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
notion-client = "*"
mistletoe = "^0.8.1"
beautifulsoup4 = "^4.10.0"
loguru = "^0.5.3"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
